pyproject.toml
[tool.poetry]
name = "cardgenerator"
version = "0.1.0"
description = "DnD Item Generator in the style of a collectible card"
authors = ["Alan Meigs <alan.meigs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gradio = "^5.6.0"
fal-client = "^0.5.6"
pillow = "^11.0.0"
openai = "^1.55.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
